Separate logger to own top level class
